Extract BufferedCancelIndicator
